services/cityflow-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
watchfiles==0.24.0
websockets>=12.0
